Refactored ContessaException
