+++ src/main/java/com/studiomediatech/contessa/store/none/NoneStorageImpl.java
@@ -3,7 +3,7 @@ package com.studiomediatech.contessa.store.none;
 import com.studiomediatech.contessa.domain.Entry;
 import com.studiomediatech.contessa.logging.Loggable;
 import com.studiomediatech.contessa.store.Storage;
-import com.studiomediatech.contessa.store.minio.ContessaException;
+import com.studiomediatech.contessa.store.ContessaException;
 
 import java.util.Optional;
